issue #2 add redux-devtools to the project. the dev tools will be auto excluded when process.env.NODE_ENV === 'production'
