+++ package.json
@@ -35,6 +35,9 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "react-hot-loader": "^1.3.0",
+    "redux-devtools": "^3.0.1",
+    "redux-devtools-dock-monitor": "^1.0.1",
+    "redux-devtools-log-monitor": "^1.0.1",
     "webpack": "^1.12.9",
     "webpack-dev-server": "^1.14.0"
   }
