+++ src/containers/Root.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./Root.prod.jsx');
+} else {
+  module.exports = require('./Root.dev.jsx');
+}
