+++ src/containers/Root.prod.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import { Provider } from 'react-redux';
+import App from './App.jsx';
+
+export default class Root extends Component {
+  render() {
+    const { store } = this.props;
+    return (
+      <Provider store={store}>
+          <App />
+      </Provider>
+    );
+  }
+}
