+++ src/index.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import configureStore from './store/configureStore';
-import App from './containers/App.jsx';
+import Root from './containers/Root';
 import {Provider} from 'react-redux';
+
 let store = configureStore();
 
 ReactDOM.render((
-  <Provider store={store}>
-    <App />
-  </Provider>
+  <Root store={store} />
 ), document.getElementById('app'));
