+++ src/store/configureStore.prod.js
@@ -1,0 +1,13 @@
+import { createStore, applyMiddleware, compose } from 'redux';
+import rootReducer from '../reducers';
+import DevTools from '../containers/DevTools.jsx';
+
+const finalCreateStore = compose(
+  // Middleware you want to use in development: no middleware is required in current.
+  // applyMiddleware(d1, d2, d3),
+  // Required! Enable Redux DevTools with the monitors you chose
+)(createStore);
+
+export default function configureStore(initialState) {
+  return finalCreateStore(rootReducer, initialState);
+}
